+++ subcommand/controller/command.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package aclcontroller
+package controller
 
 import (
 	"bytes"
@@ -402,7 +402,7 @@ func (c *Command) upsertBindingRule(consulClient *api.Client, bindingRule *api.A
 
 // upsertAnonymousTokenPolicy ensures that the anonymous ACL token has the correct permissions
 // to allow cross-DC communication via mesh gateways.
-// If the ACL controller is in the primary datacenter then we need to update the anonymous token
+// If the controller is in the primary datacenter then we need to update the anonymous token
 // with service:read and node:read.
 // Tokens are stripped from cross DC API calls so cross DC API calls use the anonymous
 // token. Mesh gateway proxies use the anonymous token to talk cross-DC and they require
