+++ subcommand/controller/command_ent_test.go
@@ -3,7 +3,7 @@
 
 //go:build enterprise
 
-package aclcontroller
+package controller
 
 import (
 	"testing"
